Add initial revision
The program can extract given pages from a given PDF file and
store them in a new PDF file. The pages to extract can be given
in typical "I tell the printer which pages to print style", so
for example:

- 1 => only the first page
- 4-8 => pages 4,5,6,7 and 8
- 1,3,6-8,10- => pages 1,3,6,7,8 and everything from 10 onwards

At least that is the idea, I have not tested everything to be honest..

But at least it won't corrupt any file systems or perform other bad
operations so I consider this a good starting point.

I should probably also mention that it is by no means fast.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>